moving things around to make more sense. main package now consumes a tcpclient. Removed version variable (not actually using it). Updating tests
